+++ spicerack/tests/unit/test_init.py
@@ -44,7 +44,7 @@ from spicerack.mysql import Mysql
 from spicerack.netbox import Netbox, NetboxServer
 from spicerack.orchestrator import Orchestrator
 from spicerack.peeringdb import PeeringDB
-from spicerack.puppet import PuppetHosts, PuppetMaster, PuppetServer
+from spicerack.puppet import PuppetHosts, PuppetServer
 from spicerack.redfish import RedfishDell, RedfishSupermicro
 from spicerack.redis_cluster import RedisCluster
 from spicerack.remote import Remote, RemoteError, RemoteHosts
@@ -158,20 +158,6 @@ def test_spicerack_icinga(mocked_command_file, mocked_remote_query, mocked_dns, 
     mocked_hostname.assert_called_once_with()
 
 
-@mock.patch("spicerack.get_puppet_ca_hostname", return_value="puppetmaster.example.com")
-@mock.patch("spicerack.remote.Remote.query", autospec=True)
-def test_spicerack_puppet_master(mocked_remote_query, mocked_get_puppet_ca_hostname):
-    """An instance of Spicerack should allow to get a PuppetMaster instance."""
-    host = mock.MagicMock(spec_set=RemoteHosts)
-    host.__len__.return_value = 1
-    mocked_remote_query.return_value = host
-    spicerack = Spicerack(verbose=True, dry_run=False, **SPICERACK_TEST_PARAMS)
-
-    assert isinstance(spicerack.puppet_master(), PuppetMaster)
-    mocked_get_puppet_ca_hostname.assert_called_once_with()
-    assert mocked_remote_query.called
-
-
 @mock.patch("spicerack.get_ca_via_srv_record", return_value="puppetserver1001.example.org")
 @mock.patch("spicerack.remote.Remote.query", autospec=True)
 def test_spicerack_puppet_server(mocked_remote_query, mocked_get_ca_via_srv_record):
@@ -207,7 +193,7 @@ def test_spicerack_management_consoles(mocked_netbox, monkeypatch, manufacturer,
     spicerack = Spicerack(verbose=True, dry_run=False, **SPICERACK_TEST_PARAMS)
 
     assert spicerack.management_password() == "env_password"
-    assert isinstance(spicerack.ipmi("test-mgmt.example.com"), Ipmi)
+    assert isinstance(spicerack.ipmi("test-mgmt.example.com", username="batman"), Ipmi)
     if manufacturer == "fancy-but-not-supported-yet":
         with pytest.raises(
             SpicerackError, match=f"The manufacturer {manufacturer} set in Netbox for test-mgmt01 is not supported."